17874: Adds check for non-numeric continuous features when 'data_type' is unspecified (#44)
If the attributes for a feature have "type"="continuous" and "data_type"
is unspecified, then we expect those values to be numeric.

Here I add a check so that we catch if the values for a feature like
this are not numeric. This was specifically raised when a user trained
strings in a continuous feature without the "data_type"="string"
attribute. Which broke a lot of stuff downstream.