20332: Fix JSON serialization of `sample`s in feature attributes (#188)
In cases where a Numpy dtype is returned for the `sample` of a feature's
attributes, we need to make sure it is JSON serializable. After some
discussion, it was agreed that this is for display only, so, a string
representation is desired.

I considered all the available datatypes here:

![Numpy dtypes
graph](https://docs.scipy.org/doc/numpy-1.13.0/_images/dtype-hierarchy.png)

Each have a solid `str()` implementation.