20725: Add multiprocessing to infer_feature_attributes for dataframes (#239)
This PR improves `infer_feature_attributes` performance for large,
dataframe-based datasets. In addition to reduced time to calculate, the
memory overhead can also be reduced. The largest performance gains and
memory reduction come from larger datasets.

For example:
A dataset with 27 columns and 69M rows:

- `0h52m27.8s / 12,781 MiB RAM` - without multiprocessing: 
- `0h09m01.5s /    793 MiB RAM` - with multiprocessing

This represents a performance _boost_ of 5.8X and a memory _reduction_
of 16X.

By default, IFR will automatically determine whether to use
multi-processing. If the product of the dataset's rows and columns is
smaller than 25M cells, IFR will automatically _not_ use
multi-processing as it is unlikely to be faster.

Multiprocessing can be forced `ON`, regardless of datasize by including
the new kwarg `max_workers` with a value of `1` or greater (take care to
not set it much higher than the available CPU cores). Old,
non-multi-processing behavior can be enforced by setting `max_workers`
to `0`. Leaving `max_workers` undefined will use the automatic behavior
described above.