23666: Removes pre-save error on Trainee create, replaces with warning (#422)
Fixes #409

While we want to leave the check in place to determine whether the
Trainee can be saved at the default persist path or CWD, some use cases
may dictate the Trainee need not be saved to the file system. Thus, upon
creating the Trainee, we should only raise a warning and save the error
until the Trainee is intentionally saved.