23677: Adjusts batch scaler to work better for heavier workloads
The existing implementation and usage of the batch scaler is not optimal for heavy
batched work loads. During batch_react, for example, if any one react case took
longer than 75 seconds, the batch scaler would scale down to 1 over time. In these
cases, the ideal work-load is one case per available CPU core where Hows Engine is
operating.

This can get complicated in environments where auto-vertical-scaling or auto-
horizontal-scaling is possible/likely. In order to correctly adapt running batch
in these environments, it is important that the scaler is up-to-date with the
number of available threads and dynamically changes to always be a multiple of
this number.

This PR implements these changes.

-