23677: Adjusts batch scaler to work better for heavier workloads (#424)
Ensure that the batch scaler is informed of the number of available
threads/cores and update the minimum batch size to this value at each
batch event. This ensures that the available compute is used
efficiently. Also, adjustments are made in the batch scaler algorithm to
ensure that batches are adjusted with precision, but that its output is
quantized to an integer multiple of the minimum batch size, again for
compute efficiency.

This can get complicated in environments where auto-vertical-scaling is
possible/likely. In order to correctly adapt running batch in these
environments, it is important that the scaler is up-to-date with the
number of available threads and dynamically changes to always be a
multiple of this number.

This PR implements these changes.

---------

Co-authored-by: howso-automation <support@howso.com>