24002: Fixes key values for non-string nominal CAPs and confusion matrices, MINOR (#457)
Extra work must be done in the client layer to clean up the CAPs maps
and confusion matrices for non-string nominal features. This is
necessary because the associative arrays returned by the Howso Engine
are formatted into JSON which does not allow for non-string keys in
objects. So here I add code that converts the appropriate feature-value
keys back into the type the user should expect given the feature.