24289: Ensures string features don't change type in a deserialized DataFrame if IFA detects they are datetimes (#484)
We have recently improved the ability of `infer_feature_attributes` to
detect if a feature in a DataFrame with an `object` `dtype` is actually
a date or datetime. However, internally, we must preserve the fact that
the feature was originally a string in the data, so that when we
deserialize the data from a `Trainee` and convert it back to a
`DataFrame`, that original type is not lost.