24384: allow batched react to run in parallel, MAJOR
If react, react_series, or react_series_stationary runs in batch mode,
and the client practically supports running multiple requests in
parallel, run each batch in a separate thread.

Reworks the batch scaler to have a cleaner interface.

No changes to the public API.  Removes the
`AbstractHowsoClient.batch_scaler_class` property, which is unlikely
to be changed in practice.