24810: Performance improvements to `infer_delta_min_max` - MINOR (#551)
Primary changes:

### Changes that benefit all ADCs
- Previously, ProcessPoolExecutors were deployed several layers deep
inside of a loop over features -- these have been removed and pulled to
the top so that the Pool is executed over the chunks of the data.
- The best chunk_size to use for chunking large datasources is now
inferred using the available cores, memory, the total size of the data
source and the average size per row. This should help it adapt for best
performance regardless of the input data.
- Uses `lazy_map` (borrowed from Enterprise), to ensure that we do not
greedily chunk-up data from the source until we have the resources to
process them, this is another measure to make best use of the available
resources.
- Sensible boundaries to ensure 20_000 <= chunk_size <= 200_000 and
sized to span evenly across available cores.
- Rather than deep-copy of the features for each chunk, a much faster
shallow copy is made -- this is all that is needed because we're only
adding top-level keys, not modifying nested ones.
  - These changes yield about 15% performance boost for all ADCs.

- Fixes **important latent bugs**, specifically,
  - `delta_max` was being set to the measured `delta_min`
  - `rate_max` was being set to the measured `rate_min`
- This will very likely lead to improved accuracy for time-series
operations

- Changed behavior
- Nominals and semi-structured continuous no longer get an assigned
`time_series:type`

### New Spark-native path for SparkDataFrameData (and potentially other,
future Spark-native data sources):
- Separate implementation that computes the `rate_min`, `rate_max`,
`delta_min`, `delta_max`, etc. with Spark-native calls distributed into
task workers.
- On the same hardware as the non-Spark path, this yields about +77%
performance for 2M rows of test data.
  - The performance gains are greater with larger data-sources greater
- Much greater performance can be seen when utilizing multi-node
clusters.
- Adds additional testing specifically for SparkNative IFA for TS and
for semi-structured.

---------

Co-authored-by: howso-automation <support@howso.com>